+++ pkg/component/gardener/discoveryserver/deployment.go
@@ -5,6 +5,8 @@
 package discoveryserver
 
 import (
+	"fmt"
+
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
@@ -82,6 +84,8 @@ func (g *gardenerDiscoveryServer) deployment(
 								"--tls-cert-file=" + volumeMountPathTLS + "/" + secretsutils.DataKeyCertificate,
 								"--tls-private-key-file=" + volumeMountPathTLS + "/" + secretsutils.DataKeyPrivateKey,
 								"--kubeconfig=" + gardenerutils.PathGenericKubeconfig,
+								fmt.Sprintf("--workload-identity-openid-configuration-file=%s/%s", volumeMountPathWorkloadIdentity, openIDConfigDataKey),
+								fmt.Sprintf("--workload-identity-jwks-file=%s/%s", volumeMountPathWorkloadIdentity, jwksDataKey),
 							},
 							Resources: corev1.ResourceRequirements{
 								Requests: map[corev1.ResourceName]resource.Quantity{
@@ -138,6 +142,12 @@ func (g *gardenerDiscoveryServer) deployment(
 								{
 									Name:      volumeNameTLS,
 									MountPath: volumeMountPathTLS,
+									ReadOnly:  true,
+								},
+								{
+									Name:      volumeNameWorkloadIdentity,
+									MountPath: volumeMountPathWorkloadIdentity,
+									ReadOnly:  true,
 								},
 							},
 						},
@@ -147,7 +157,17 @@ func (g *gardenerDiscoveryServer) deployment(
 							Name: volumeNameTLS,
 							VolumeSource: corev1.VolumeSource{
 								Secret: &corev1.SecretVolumeSource{
-									SecretName: secretNameTLS,
+									SecretName:  secretNameTLS,
+									DefaultMode: ptr.To[int32](0400),
+								},
+							},
+						},
+						{
+							Name: volumeNameWorkloadIdentity,
+							VolumeSource: corev1.VolumeSource{
+								Secret: &corev1.SecretVolumeSource{
+									SecretName:  secretNameWorkloadIdentity,
+									DefaultMode: ptr.To[int32](0400),
 								},
 							},
 						},
@@ -157,39 +177,8 @@ func (g *gardenerDiscoveryServer) deployment(
 		},
 	}
 
-	injectWorkloadIdentity(deployment, secretNameWorkloadIdentity)
 	utilruntime.Must(gardenerutils.InjectGenericKubeconfig(deployment, secretNameGenericTokenKubeconfig, secretNameVirtualGardenAccess))
 	utilruntime.Must(references.InjectAnnotations(deployment))
 
 	return deployment
-}
-
-func injectWorkloadIdentity(deployment *appsv1.Deployment, secretName string) {
-	deployment.Spec.Template.Spec.Containers[0].Args = append(
-		deployment.Spec.Template.Spec.Containers[0].Args,
-		"--workload-identity-openid-configuration-file="+volumeMountPathWorkloadIdentity+"/"+openIDConfigDataKey,
-		"--workload-identity-jwks-file="+volumeMountPathWorkloadIdentity+"/"+jwksDataKey,
-	)
-
-	deployment.Spec.Template.Spec.Containers[0].VolumeMounts = append(
-		deployment.Spec.Template.Spec.Containers[0].VolumeMounts,
-		corev1.VolumeMount{
-			Name:      volumeNameWorkloadIdentity,
-			MountPath: volumeMountPathWorkloadIdentity,
-			ReadOnly:  true,
-		},
-	)
-
-	deployment.Spec.Template.Spec.Volumes = append(
-		deployment.Spec.Template.Spec.Volumes,
-		corev1.Volume{
-			Name: volumeNameWorkloadIdentity,
-			VolumeSource: corev1.VolumeSource{
-				Secret: &corev1.SecretVolumeSource{
-					SecretName:  secretName,
-					DefaultMode: ptr.To[int32](0400),
-				},
-			},
-		},
-	)
 }
