+++ pkg/utils/workloadidentity/openidmeta_test.go
@@ -5,6 +5,9 @@
 package workloadidentity_test
 
 import (
+	"crypto/ecdh"
+	"crypto/ecdsa"
+	"crypto/elliptic"
 	"crypto/rand"
 	"crypto/rsa"
 	"encoding/json"
@@ -20,7 +23,7 @@ var _ = Describe("#OpenIDMeta", func() {
 	Describe("#OpenIDConfig", func() {
 		It("should correctly build the openid configuration document", func() {
 			const issuerURL = "https://test.gardener.cloud/workload-identity"
-			key, err := rsa.GenerateKey(rand.Reader, 2048)
+			key, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
 			Expect(err).ToNot(HaveOccurred())
 
 			openedIDConfig, err := workloadidentity.OpenIDConfig(issuerURL, key.Public())
@@ -31,7 +34,7 @@ var _ = Describe("#OpenIDMeta", func() {
 				JWKSURI:       issuerURL + "/jwks",
 				ResponseTypes: []string{"id_token"},
 				SubjectTypes:  []string{"public"},
-				SigningAlgs:   []string{"RS256"},
+				SigningAlgs:   []string{"ES256"},
 			}
 			expectedOpenIDConfigMarshaled, err := json.Marshal(expectedOpenIDConfig)
 			Expect(err).ToNot(HaveOccurred())
@@ -45,7 +48,7 @@ var _ = Describe("#OpenIDMeta", func() {
 			Expect(err).ToNot(HaveOccurred())
 
 			jwks, err := workloadidentity.JWKS(key)
-			Expect(err).To(MatchError("JWKS supports only public keys"))
+			Expect(err).To(MatchError("all keys must be public"))
 			Expect(jwks).To(BeNil())
 		})
 
@@ -76,6 +79,47 @@ var _ = Describe("#OpenIDMeta", func() {
 			Expect(err).ToNot(HaveOccurred())
 			Expect(jwks).To(Equal(expectedJWKSMarshaled))
 		})
+	})
 
+	Describe("#getAlg", func() {
+		It("should get correct algorithm", func() {
+			privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
+			Expect(err).ToNot(HaveOccurred())
+			alg, err := workloadidentity.GetAlg(privateKey.Public())
+			Expect(err).ToNot(HaveOccurred())
+			Expect(alg).To(Equal(jose.RS256))
+
+			ecdsaPrivateKey, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
+			Expect(err).ToNot(HaveOccurred())
+			alg, err = workloadidentity.GetAlg(ecdsaPrivateKey.Public())
+			Expect(err).ToNot(HaveOccurred())
+			Expect(alg).To(Equal(jose.ES256))
+
+			ecdsaPrivateKey, err = ecdsa.GenerateKey(elliptic.P384(), rand.Reader)
+			Expect(err).ToNot(HaveOccurred())
+			alg, err = workloadidentity.GetAlg(ecdsaPrivateKey.Public())
+			Expect(err).ToNot(HaveOccurred())
+			Expect(alg).To(Equal(jose.ES384))
+
+			ecdsaPrivateKey, err = ecdsa.GenerateKey(elliptic.P521(), rand.Reader)
+			Expect(err).ToNot(HaveOccurred())
+			alg, err = workloadidentity.GetAlg(ecdsaPrivateKey.Public())
+			Expect(err).ToNot(HaveOccurred())
+			Expect(alg).To(Equal(jose.ES512))
+		})
+
+		It("should return error for unknown ECDSA curve", func() {
+			key, err := ecdsa.GenerateKey(elliptic.P224(), rand.Reader)
+			Expect(err).ToNot(HaveOccurred())
+			_, err = workloadidentity.GetAlg(key.Public())
+			Expect(err).To(MatchError("unknown ecdsa key curve, must be 256, 384, or 521"))
+		})
+
+		It("should return error for unknown key type", func() {
+			key, err := ecdh.P256().GenerateKey(rand.Reader)
+			Expect(err).ToNot(HaveOccurred())
+			_, err = workloadidentity.GetAlg(key.Public())
+			Expect(err).To(MatchError("unknown public key type, must be *rsa.PublicKey, *ecdsa.PublicKey, or jose.OpaqueSigner"))
+		})
 	})
 })
